refactor: change users/views.py
